slide_gen/requirements.txt
gradio
groq==0.7.0
huggingface_hub
numpy
Pillow
pydantic
python-dotenv
python_pptx
Requests
scipy
setuptools
scikit-image
torch
torchvision
tqdm
uvicorn
httpx==0.23.3
einops
accelerate
sentencepiece